Add accessibility for blind users.
Added accessibility for blind users. Talkback can't handle FontAwesome or the CodeEdit control. Added a checkbox in settings, and modified the program to use accessible text/controls. ***UNTESTED*** I hope this will be integrated in to the official droidscript app, but I cannot test it. I have created this as a starting point, hoping for more help and refinements from the droidscript team.